test: restructure
