+++ tests/fixtures/cpp/sample.cpp
