+++ tests/fixtures/graphql/sample.graphql
