+++ tests/fixtures/javascript/basic.js
@@ -1,0 +1,141 @@
+import {
+  a,
+  b,
+  MARKER
+  c
+} from "mod";
+
+export {
+  a,
+  b,
+  MARKER
+}
+
+foo({
+    sd,
+    aaa: sdf[
+      "index" +
+        2
+    ]
+  },
+  4
+);
+
+foo(2, {
+    sd,
+    sdf,
+    MARKER
+  },
+  4
+);
+
+foo( 2,
+  {
+    sd,
+    sdf
+  });
+
+foo(2, {
+  sd,
+  sdf
+});
+
+foo(2, {
+  sd,
+  sdf
+}, 'abc');
+
+foo(2,
+  {
+    sd,
+    sdf
+  },
+  'abc');
+
+foo(2,
+  4);
+
+var x = [
+  3,
+  4
+];
+
+const y = [
+  1
+];
+
+const j = [{
+  a: 1
+}];
+
+let h = {
+  a: [ 1,
+    MARKER
+    2 ],
+  b: { j: [
+    { l: 1 }]
+  },
+  c:
+    { j: [
+      MARKER
+      { l: 1 }]
+    },
+};
+
+const a =
+  {
+    b: 1
+  };
+
+function func(
+  fdsf
+) {
+  const f = () => {
+    MARKER
+    fs
+  }
+}
+
+const af1 = (c) => {
+  const a = () => {
+    MARKER
+    sdf
+  }
+}
+
+const af2 = (c,
+  b,
+  d
+) => ({
+  f: () => {
+    MARKER
+    sdf
+  }
+})
+
+class MyClass extends OtherComponent {
+
+  state = {
+    test: 1
+  }
+
+  constructor() {
+    test();
+  }
+  MARKER
+
+  otherfunction = (a, b = {
+    default: false
+  }) => {
+    more();
+  }
+}
+
+foo(myWrapper(mysecondWrapper({
+  /**
+   * Comment
+   * Comment
+   */
+  a: 1
+})));
+
