+++ tests/fixtures/javascript/binary.js
@@ -1,0 +1,33 @@
+b =
+  3
+    + 5
+    + 7
+    + 8
+      * 8
+      * 9
+      / 17
+      * 8
+      / 20
+    - 34
+    + 3 *
+      9
+    - 8;
+
+ifthis
+  && thendo()
+  && aaaa
+  || (otherwise
+    && dothis
+    && bbbb
+    && cccc) &&
+    fff &&
+    fds ||
+  (fffff &&
+    foo(
+      a,
+      b,
+      MARKER
+    ) &&
+    aaa)
+
+
