+++ tests/fixtures/javascript/chained_call.js
@@ -1,0 +1,8 @@
+req
+  .field
+  MARKER
+  .shouldBeOne()
+  .abc(function() {
+    MARKER
+  })
+
