+++ tests/fixtures/javascript/iife.js
@@ -1,0 +1,20 @@
+function foo() {
+  return (function(
+    a,
+    b
+  ) {
+    MARKER
+    ff
+  })(
+    c
+    MARKER
+  )(
+    d
+    MARKER
+  )(123, function () {
+    dosome()
+  }, {
+    a: 456
+  });
+}
+
