+++ tests/fixtures/javascript/jsx.js
@@ -1,0 +1,54 @@
+const jsx = (
+  <div
+    title='start'
+    ff={
+      ((
+        fd
+      ) => {
+        fd
+        MARKER
+      })()
+    }
+    MARKER
+  >
+    <div>
+      MARKER
+      sdf
+    </div>
+    <div>
+      {isA? (
+        <div>
+          aaaaaaaaaaaaaaa
+        </div>) : (<span>
+          fsdfsff
+          MARKER
+        </span>)
+      }
+      {isA? (
+        <div>
+          aaaaaaaaaaaaaaa
+        </div>
+      ) : (
+          <span>
+            {/*
+             * FIXME: fix this
+             */}
+            fsdfsff
+            MARKER
+          </span>
+        )}
+    </div>
+  </div>
+);
+
+const a = (
+  <img
+    src='/img.jpg'
+  />
+);
+
+const b = (
+  <img
+    src='/img.jpg' />
+);
+
