+++ tests/fixtures/lua/sample.lua
