+++ tests/fixtures/tsx/sample.tsx
@@ -1,0 +1,14 @@
+export function List() {
+  return (
+    <List>
+      <div>
+        {
+          MARKER
+        }
+      </div>
+      {
+        MARKER
+      }
+    </List>
+  )
+}
