+++ tests/fixtures/typescript/return_type.ts
@@ -1,0 +1,10 @@
+function func(
+  aaa: number,
+  bbb: boolean
+): {
+  ac: T; // TODO: extra indent
+  pppp: P;
+  MARKER
+} {
+  MARKER
+}
