+++ tests/fixtures/vue/sample.vue
