+++ tests/install.vim
