+++ tests/preload.vim
